Use uint8 array for point sizes and do different round circles for node vs webgl

+++ js/loadgml.js
@@ -54,7 +54,7 @@ function loadGMLJSON (path) {
             var res = {
                 nodes: {
                     colors: new Uint32Array(_.pluck(nodes, 'color')),
-                    sizes: new Float32Array(nodes.length)
+                    sizes: new Uint8Array(nodes.length)
                 },
                 edges: data.links.map(function (o) { return [o.source, o.target]; }),
                 //BUG some reason
